requirements.txt
plotext>=5.2.8
prometheus-api-client>=0.5.3
requests>=2.28.0
click>=8.0.0
PyYAML>=6.0